refactor: move prisma to server folder
